Complete production role workflows

+++ QuickBite/src/main/java/com/debbiecyber/QuickBite/dto/response/CartItemResponse.java
@@ -3,6 +3,7 @@ package com.debbiecyber.QuickBite.dto.response;
 
 import lombok.Builder;
 import lombok.Data;
+import java.math.BigDecimal;
 
 @Data
 @Builder
@@ -17,11 +18,11 @@ public class CartItemResponse {
 
     private String itemName;
 
-    private Double price;
+    private BigDecimal price;
 
     private Integer quantity;
 
-    private Double subTotal;
+    private BigDecimal subTotal;
 
     private String imageURL;
