+++ QuickBite/src/main/java/com/debbiecyber/QuickBite/dto/response/OrderItemResponse.java
@@ -3,6 +3,7 @@ package com.debbiecyber.QuickBite.dto.response;
 
 import lombok.Builder;
 import lombok.Data;
+import java.math.BigDecimal;
 
 @Data
 @Builder
@@ -11,9 +12,9 @@ public class OrderItemResponse {
 
     private String itemName;
 
-    private Double price;
+    private BigDecimal price;
 
     private Integer quantity;
 
-    private Double subTotal;
+    private BigDecimal subTotal;
 }
